+++ R/AllGenerics.R
@@ -16,8 +16,8 @@
 #' # Example data ----
 #' library(GeneSet)
 #' tgs <- tbl_geneset(
-#'     "Cell type 1" = c("Gene001", "Gene002"),
-#'     "Cell type 2" = c("Gene002", "Gene003", "Gene004")
+#'     "Cell type 1"=c("Gene001", "Gene002"),
+#'     "Cell type 2"=c("Gene002", "Gene003", "Gene004")
 #' )
 #'
 #' # Example usage ----
@@ -49,8 +49,8 @@ setGeneric(
 #' # Example data ----
 #' library(GeneSet)
 #' tgs <- tbl_geneset(
-#'     "Cell type 1" = c("Gene001", "Gene002"),
-#'     "Cell type 2" = c("Gene002", "Gene003", "Gene004")
+#'     "Cell type 1"=c("Gene001", "Gene002"),
+#'     "Cell type 2"=c("Gene002", "Gene003", "Gene004")
 #' )
 #'
 #' # Example usage ----