+++ tests/testthat/test-GeneSet.R
@@ -2,8 +2,8 @@
 # Example data ----
 
 tgs <- tbl_geneset(
-    "Cell type 1" = c("Gene001", "Gene002"),
-    "Cell type 2" = c("Gene002", "Gene003", "Gene004")
+    "Cell type 1"=c("Gene001", "Gene002"),
+    "Cell type 2"=c("Gene002", "Gene003", "Gene004")
 )
 
 # .makeFilterExpressionFromGeneSetCollection ----