+++ tests/testthat/test-predict.R
@@ -13,14 +13,14 @@ gsc <- GeneSetCollection(list(
 ))
 
 tgs <- tbl_geneset(
-    "Cell type 1" = c("Gene001", "Gene002"),
-    "Cell type 2" = c("Gene002", "Gene003", "Gene004")
+    "Cell type 1"=c("Gene001", "Gene002"),
+    "Cell type 2"=c("Gene002", "Gene003", "Gene004")
 )
 
 # predict.GeneSetCollection ----
 
 test_that("predict.GeneSetCollection works for method ProportionPositive", {
-    dummyCluster <- factor(sample(head(LETTERS, 3), ncol(se), replace = TRUE))
+    dummyCluster <- factor(sample(head(LETTERS, 3), ncol(se), replace=TRUE))
     colData(se)[, "cluster"] <- dummyCluster
 
     out <- predict(gsc, se, method="ProportionPositive", cluster.col="cluster")
@@ -56,7 +56,7 @@ test_that("predict.GeneSetCollection works for method ProportionPositive", {
 })
 
 test_that("predict.tbl_geneset works for method ProportionPositive", {
-    dummyCluster <- factor(sample(head(LETTERS, 3), ncol(se), replace = TRUE))
+    dummyCluster <- factor(sample(head(LETTERS, 3), ncol(se), replace=TRUE))
     colData(se)[, "cluster"] <- dummyCluster
 
     out <- predict(tgs, se, method="ProportionPositive", cluster.col="cluster")
@@ -91,14 +91,14 @@ test_that("predict.tbl_geneset works for method ProportionPositive", {
     expect_s4_class(plotOut, "Heatmap")
 })
 
-test_that("predictProportionSignatureByCluster requires argument col.cluster", {
+test_that("predictByProportionPositive requires argument col.cluster", {
     expect_error(
-        predictProportionSignatureByCluster(gsc, se),
+        predictByProportionPositive(gsc, se),
         "cluster.col is required for method 'ProportionPositive'"
     )
 })
 
-test_that("plotProportionPositive requires predictProportionSignatureByCluster results", {
+test_that("plotProportionPositive requires predictByProportionPositive results", {
     expect_error(
         plotProportionPositive(se),
         "Method 'ProportionPositive' was not run yet.")